added app tests for checkout service
