--- source/API/checkout_api.py
@@ -1,81 +1,0 @@
-# from flask import Flask, request, jsonify
-# from typing import Any, Optional
-#
-# # Handle missing imports with mocks
-# try:
-#     from source.services.cart_manager import CartManager
-#     from source.services.inventory_manager import InventoryManager
-#     from source.services.order_manager import OrderManager
-#     from source.services.user_manager import UserManager
-#     from source.services.checkout_service import CheckoutService
-# except ImportError:
-#
-#     class CartManager:
-#         def get_cart(self, user_id: int):
-#             return None  # Mock response
-#
-#     class InventoryManager:
-#         def is_item_available(self, item_id: int, quantity: int) -> bool:
-#             return True  # Mock response
-#
-#     class OrderManager:
-#         def get_order(self, order_id: int):
-#             return None  # Mock response
-#
-#     class UserManager:
-#         def get_user(self, user_id: int):
-#             return None  # Mock response
-#
-#     class CheckoutService:
-#         def __init__(self, cart_manager, inventory_manager, order_manager, user_manager):
-#             self.cart_manager = cart_manager
-#             self.inventory_manager = inventory_manager
-#             self.order_manager = order_manager
-#             self.user_manager = user_manager
-#
-#         def finalize_checkout(self, user_id: int, address: str, payment_method: str):
-#             return {"status": "error", "message": "Mock CheckoutService - Not Implemented"}
-#
-#
-# # Initialize services
-# cart_manager = CartManager()
-# inventory_manager = InventoryManager()
-# order_manager = OrderManager()
-# user_manager = UserManager()
-# checkout_service = CheckoutService(cart_manager, inventory_manager, order_manager, user_manager)
-#
-#
-# # Flask app factory
-# def create_app() -> Flask:
-#     """Creates and configures the Flask application."""
-#     app = Flask(__name__)
-#
-#     @app.route('/checkout', methods=['POST'])
-#     def start_checkout() -> Any:
-#         """Handles the checkout process."""
-#         data = request.get_json(silent=True)
-#         if not data:
-#             return jsonify({"status": "error", "message": "Invalid JSON format"}), 400
-#
-#         user_id: Optional[int] = data.get("user_id")
-#         address: Optional[str] = data.get("address")
-#         payment_method: Optional[str] = data.get("payment_method")
-#
-#         if user_id is None or not address or not payment_method:
-#             return jsonify({"status": "error", "message": "Missing required fields"}), 400
-#
-#         # Handle invalid payment method
-#         valid_payment_methods = {"credit_card", "paypal", "bank_transfer"}
-#         if payment_method not in valid_payment_methods:
-#             return jsonify({"status": "error", "message": "Invalid payment method"}), 400
-#
-#         result = checkout_service.finalize_checkout(user_id, address, payment_method)
-#         return jsonify(result), 200 if result["status"] == "success" else 400
-#
-#     return app
-#
-#
-# # Run Flask app
-# if __name__ == '__main__':
-#     app = create_app()
-#     app.run(debug=True)
